circle_test_upgrade

+++ stylesheets/db_upgrade.sql
@@ -89,7 +89,7 @@ IF NOT EXISTS
 	(SELECT * FROM information_schema.columns 
 		WHERE table_schema='spark' AND table_name ='categories' AND column_name ='courseAreaId') 
 THEN 
-	ALTER TABLE categories ADD courseAreaId TINYINT(1) DEFAULT 0;
+	ALTER TABLE categories ADD courseAreaId INT DEFAULT 0;
 END IF;
 
 END;
